Output `Rmd` file to html.  The table formatting is frustratingly inconsistent between sections in the document.
The output includes a test-case of the run, for dataset 1001, `1001_test.csv`.  This is a more complete output than last time (archived in the Neotoma Google Docs).  Some fine points need to be addressed still.  The actual document needs to be converted to a function, and likely to a SQL Server stored procedure.

There are sets of relative ages that are not cross-referenced (see the `NDB.relativeages` table), but many ages don't have any assigned date either.