Added `dplyr` to the `functional_run`.
Added some messages for debugging.

Added a skip for datasets without biological content.

Reformatted the Query string so it looks a bit cleaner.

Moved the contact string to a new query, there are cases where multiple contacts are coming in and it's causing the size of the output to double.

Added some blocks to clean out dataset types that aren't biological.  Need to add`UPA`.

Added new block to manage multiple default age types assigned as default.

Moved actual ages into a dynamic properties field.

Had to correct some case-sensitive fields in `query_out` when assigning to `output`.

Added a block for `identificationQualifiers`

There are changes in the assignment for geopolitical sites, but they're mostly cosmetic.