Added flow chart images for the `Rmd`.
I added macroinvertebrates back in.  They're down as Mollusca, that's fine.

Cleared the "cooccurrs with" fields when a sample is a signleton.

Added clearer references to packages in the function calls.

Ran the partial `Rmd` file, so I could get something more complete done.

Added the SVG and PNG file to the `Rmd`.  Added a block of bullet points for the actually transcribed fields.  Left the tables & table descriptions below.  Added in field descriptions for most of the DarwinCore fields described.  Left some blank when I got lazy or felt they were pretty descriptive already.